chore: rename to canvasclient pkg
